Struct rp2040_pac2::adc::regs::Fcs

#[repr(transparent)]pub struct Fcs(pub u32);

FIFO control and status

Implementations

impl Fcs

pub const fn thresh(&self) -> u8

DREQ/IRQ asserted when level >= threshold

pub fn set_thresh(&mut self, val: u8)

DREQ/IRQ asserted when level >= threshold

pub const fn level(&self) -> u8

The number of conversion results currently waiting in the FIFO

pub fn set_level(&mut self, val: u8)

The number of conversion results currently waiting in the FIFO

pub const fn over(&self) -> bool

1 if the FIFO has been overflowed. Write 1 to clear.

pub fn set_over(&mut self, val: bool)

1 if the FIFO has been overflowed. Write 1 to clear.

pub const fn under(&self) -> bool

1 if the FIFO has been underflowed. Write 1 to clear.

pub fn set_under(&mut self, val: bool)

1 if the FIFO has been underflowed. Write 1 to clear.

pub const fn full(&self) -> bool

pub fn set_full(&mut self, val: bool)

pub const fn empty(&self) -> bool

pub fn set_empty(&mut self, val: bool)

pub const fn dreq_en(&self) -> bool

If 1: assert DMA requests when FIFO contains data

pub fn set_dreq_en(&mut self, val: bool)

If 1: assert DMA requests when FIFO contains data

pub const fn err(&self) -> bool

If 1: conversion error bit appears in the FIFO alongside the result

pub fn set_err(&mut self, val: bool)

If 1: conversion error bit appears in the FIFO alongside the result

pub const fn shift(&self) -> bool

If 1: FIFO results are right-shifted to be one byte in size. Enables DMA to byte buffers.

pub fn set_shift(&mut self, val: bool)

If 1: FIFO results are right-shifted to be one byte in size. Enables DMA to byte buffers.

pub const fn en(&self) -> bool

If 1: write result to the FIFO after each conversion.

pub fn set_en(&mut self, val: bool)

If 1: write result to the FIFO after each conversion.