Struct rp2040_pac2::ppb::regs::SystCsr

#[repr(transparent)]pub struct SystCsr(pub u32);

Use the SysTick Control and Status Register to enable the SysTick features.

Implementations

impl SystCsr

pub const fn countflag(&self) -> bool

Returns 1 if timer counted to 0 since last time this was read. Clears on read by application or debugger.

pub fn set_countflag(&mut self, val: bool)

Returns 1 if timer counted to 0 since last time this was read. Clears on read by application or debugger.

pub const fn clksource(&self) -> bool

SysTick clock source. Always reads as one if SYST_CALIB reports NOREF. Selects the SysTick timer clock source: 0 = External reference clock. 1 = Processor clock.

pub fn set_clksource(&mut self, val: bool)

SysTick clock source. Always reads as one if SYST_CALIB reports NOREF. Selects the SysTick timer clock source: 0 = External reference clock. 1 = Processor clock.

pub const fn tickint(&self) -> bool

Enables SysTick exception request: 0 = Counting down to zero does not assert the SysTick exception request. 1 = Counting down to zero to asserts the SysTick exception request.

pub fn set_tickint(&mut self, val: bool)

Enables SysTick exception request: 0 = Counting down to zero does not assert the SysTick exception request. 1 = Counting down to zero to asserts the SysTick exception request.

pub const fn enable(&self) -> bool

Enable SysTick counter: 0 = Counter disabled. 1 = Counter enabled.

pub fn set_enable(&mut self, val: bool)

Enable SysTick counter: 0 = Counter disabled. 1 = Counter enabled.