Module rp2040_pac2::rosc::regs

Structs

Count
A down counter running at the ROSC frequency which counts to zero and stops. To start the counter write a non-zero value. Can be used for short software pauses when setting up time sensitive hardware.

Ctrl
Ring Oscillator control

Div
Controls the output divider

Freqa
The FREQA & FREQB registers control the frequency by controlling the drive strength of each stage The drive strength has 4 levels determined by the number of bits set Increasing the number of bits set increases the drive strength and increases the oscillation frequency 0 bits set is the default drive strength 1 bit set doubles the drive strength 2 bits set triples drive strength 3 bits set quadruples drive strength

Freqb
For a detailed description see freqa register

Phase
Controls the phase shifted output

Randombit
This just reads the state of the oscillator output so randomness is compromised if the ring oscillator is stopped or run at a harmonic of the bus frequency

Status
Ring Oscillator Status