Struct rp2040_pac2::sio::regs::Interp0CtrlLane0

#[repr(transparent)]pub struct Interp0CtrlLane0(_);

Control register for lane 0

Implementations

impl Interp0CtrlLane0

pub const fn to_bits(&self) -> u32

pub const fn from_bits(val: u32) -> Interp0CtrlLane0

pub const fn overf(&self) -> bool

Set if either OVERF0 or OVERF1 is set.

pub fn set_overf(&mut self, val: bool)

Set if either OVERF0 or OVERF1 is set.

pub const fn overf1(&self) -> bool

Indicates if any masked-off MSBs in ACCUM1 are set.

pub fn set_overf1(&mut self, val: bool)

Indicates if any masked-off MSBs in ACCUM1 are set.

pub const fn overf0(&self) -> bool

Indicates if any masked-off MSBs in ACCUM0 are set.

pub fn set_overf0(&mut self, val: bool)

Indicates if any masked-off MSBs in ACCUM0 are set.

pub const fn blend(&self) -> bool

Only present on INTERP0 on each core. If BLEND mode is enabled: - LANE1 result is a linear interpolation between BASE0 and BASE1, controlled by the 8 LSBs of lane 1 shift and mask value (a fractional number between 0 and 255/256ths) - LANE0 result does not have BASE0 added (yields only the 8 LSBs of lane 1 shift+mask value) - FULL result does not have lane 1 shift+mask value added (BASE2 + lane 0 shift+mask) LANE1 SIGNED flag controls whether the interpolation is signed or unsigned.

pub fn set_blend(&mut self, val: bool)

Only present on INTERP0 on each core. If BLEND mode is enabled: - LANE1 result is a linear interpolation between BASE0 and BASE1, controlled by the 8 LSBs of lane 1 shift and mask value (a fractional number between 0 and 255/256ths) - LANE0 result does not have BASE0 added (yields only the 8 LSBs of lane 1 shift+mask value) - FULL result does not have lane 1 shift+mask value added (BASE2 + lane 0 shift+mask) LANE1 SIGNED flag controls whether the interpolation is signed or unsigned.

pub const fn force_msb(&self) -> u8

ORed into bits 29:28 of the lane result presented to the processor on the bus. No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence of pointers into flash or SRAM.

pub fn set_force_msb(&mut self, val: u8)

ORed into bits 29:28 of the lane result presented to the processor on the bus. No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence of pointers into flash or SRAM.

pub const fn add_raw(&self) -> bool

If 1, mask + shift is bypassed for LANE0 result. This does not affect FULL result.

pub fn set_add_raw(&mut self, val: bool)

If 1, mask + shift is bypassed for LANE0 result. This does not affect FULL result.

pub const fn cross_result(&self) -> bool

If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub fn set_cross_result(&mut self, val: bool)

If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub const fn cross_input(&self) -> bool

If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware. Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub fn set_cross_input(&mut self, val: bool)

If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware. Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub const fn signed(&self) -> bool

If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits before adding to BASE0, and LANE0 PEEK/POP appear extended to 32 bits when read by processor.

pub fn set_signed(&mut self, val: bool)

If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits before adding to BASE0, and LANE0 PEEK/POP appear extended to 32 bits when read by processor.

pub const fn mask_msb(&self) -> u8

The most-significant bit allowed to pass by the mask (inclusive) Setting MSB < LSB may cause chip to turn inside-out

pub fn set_mask_msb(&mut self, val: u8)

The most-significant bit allowed to pass by the mask (inclusive) Setting MSB < LSB may cause chip to turn inside-out

pub const fn mask_lsb(&self) -> u8

The least-significant bit allowed to pass by the mask (inclusive)

pub fn set_mask_lsb(&mut self, val: u8)

The least-significant bit allowed to pass by the mask (inclusive)

pub const fn shift(&self) -> u8

Logical right-shift applied to accumulator before masking

pub fn set_shift(&mut self, val: u8)

Logical right-shift applied to accumulator before masking