Struct rp2040_pac2::sio::regs::Interp1CtrlLane0

#[repr(transparent)]
pub struct Interp1CtrlLane0(pub u32);

Control register for lane 0

Tuple Fields
0: u32
Implementations

impl Interp1CtrlLane0

pub const fn shift(&self) -> u8

Logical right-shift applied to accumulator before masking

pub fn set_shift(&mut self, val: u8)

Logical right-shift applied to accumulator before masking

pub const fn mask_lsb(&self) -> u8

The least-significant bit allowed to pass by the mask (inclusive)

pub fn set_mask_lsb(&mut self, val: u8)

The least-significant bit allowed to pass by the mask (inclusive)

pub const fn mask_msb(&self) -> u8

The most-significant bit allowed to pass by the mask (inclusive) Setting MSB < LSB may cause chip to turn inside-out

pub fn set_mask_msb(&mut self, val: u8)

The most-significant bit allowed to pass by the mask (inclusive) Setting MSB < LSB may cause chip to turn inside-out

pub const fn signed(&self) -> bool

If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits before adding to BASE0, and LANE0 PEEK/POP appear extended to 32 bits when read by processor.

pub fn set_signed(&mut self, val: bool)

If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits before adding to BASE0, and LANE0 PEEK/POP appear extended to 32 bits when read by processor.

pub const fn cross_input(&self) -> bool

If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware. Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub fn set_cross_input(&mut self, val: bool)

If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware. Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub const fn cross_result(&self) -> bool

If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub fn set_cross_result(&mut self, val: bool)

If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub const fn add_raw(&self) -> bool

If 1, mask + shift is bypassed for LANE0 result. This does not affect FULL result.

pub fn set_add_raw(&mut self, val: bool)

If 1, mask + shift is bypassed for LANE0 result. This does not affect FULL result.

pub const fn force_msb(&self) -> u8

ORed into bits 29:28 of the lane result presented to the processor on the bus. No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence of pointers into flash or SRAM.

pub fn set_force_msb(&mut self, val: u8)

ORed into bits 29:28 of the lane result presented to the processor on the bus. No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence of pointers into flash or SRAM.

pub const fn clamp(&self) -> bool

Only present on INTERP1 on each core. If CLAMP mode is enabled: - LANE0 result is shifted and masked ACCUM0, clamped by a lower bound of BASE0 and an upper bound of BASE1. - Signedness of these comparisons is determined by LANE0_CTRL_SIGNED

pub fn set_clamp(&mut self, val: bool)

Only present on INTERP1 on each core. If CLAMP mode is enabled: - LANE0 result is shifted and masked ACCUM0, clamped by a lower bound of BASE0 and an upper bound of BASE1. - Signedness of these comparisons is determined by LANE0_CTRL_SIGNED

pub const fn overf0(&self) -> bool

Indicates if any masked-off MSBs in ACCUM0 are set.

pub fn set_overf0(&mut self, val: bool)

Indicates if any masked-off MSBs in ACCUM0 are set.

pub const fn overf1(&self) -> bool

Indicates if any masked-off MSBs in ACCUM1 are set.

pub fn set_overf1(&mut self, val: bool)

Indicates if any masked-off MSBs in ACCUM1 are set.

pub const fn overf(&self) -> bool

Set if either OVERF0 or OVERF1 is set.

pub fn set_overf(&mut self, val: bool)

Set if either OVERF0 or OVERF1 is set.