Struct rp2040_pac2::sio::Sio

pub struct Sio(pub *mut u8);

Single-cycle IO block Provides core-local and inter-core hardware for the two processors, with single-cycle access.

Tuple Fields
0: *mut u8
Implementations

impl Sio

pub fn cpuid(self) -> Reg<u32, R>

Processor core identifier Value is 0 when read from processor core 0, and 1 when read from processor core 1.

pub fn spinlock_st(self) -> Reg<u32, R>

Spinlock state A bitmap containing the state of all 32 spinlocks (1=locked). Mainly intended for debugging.

pub fn div(self) -> Div

pub fn fifo(self) -> Fifo

pub fn interp(self, n: usize) -> Interp

pub fn spinlock(self, n: usize) -> Reg<u32, R>

Reading from a spinlock address will: - Return 0 if lock is already locked - Otherwise return nonzero, and simultaneously claim the lock Writing (any value) releases the lock. If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins. The value returned on success is 0x1 << lock number.

pub fn gpio_out(self, n: usize) -> Gpio

pub fn gpio_oe(self, n: usize) -> Gpio

pub fn gpio_in(self, n: usize) -> Reg<u32, RW>

Input value for GPIO pins