Struct rp2040_pac2::usb::regs::BuffStatus

#[repr(transparent)]
pub struct BuffStatus(pub u32);

Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

Tuple Fields
0: u32
Implementations

impl BuffStatus

pub const fn ep0_in(&self) -> bool

pub fn set_ep0_in(&mut self, val: bool)

pub const fn ep0_out(&self) -> bool

pub fn set_ep0_out(&mut self, val: bool)

pub const fn ep1_in(&self) -> bool

pub fn set_ep1_in(&mut self, val: bool)

pub const fn ep1_out(&self) -> bool

pub fn set_ep1_out(&mut self, val: bool)

pub const fn ep2_in(&self) -> bool

pub fn set_ep2_in(&mut self, val: bool)

pub const fn ep2_out(&self) -> bool

pub fn set_ep2_out(&mut self, val: bool)

pub const fn ep3_in(&self) -> bool

pub fn set_ep3_in(&mut self, val: bool)

pub const fn ep3_out(&self) -> bool

pub fn set_ep3_out(&mut self, val: bool)

pub const fn ep4_in(&self) -> bool

pub fn set_ep4_in(&mut self, val: bool)

pub const fn ep4_out(&self) -> bool

pub fn set_ep4_out(&mut self, val: bool)

pub const fn ep5_in(&self) -> bool

pub fn set_ep5_in(&mut self, val: bool)

pub const fn ep5_out(&self) -> bool

pub fn set_ep5_out(&mut self, val: bool)

pub const fn ep6_in(&self) -> bool

pub fn set_ep6_in(&mut self, val: bool)

pub const fn ep6_out(&self) -> bool

pub fn set_ep6_out(&mut self, val: bool)

pub const fn ep7_in(&self) -> bool

pub fn set_ep7_in(&mut self, val: bool)

pub const fn ep7_out(&self) -> bool

pub fn set_ep7_out(&mut self, val: bool)

pub const fn ep8_in(&self) -> bool

pub fn set_ep8_in(&mut self, val: bool)

pub const fn ep8_out(&self) -> bool

pub fn set_ep8_out(&mut self, val: bool)

pub const fn ep9_in(&self) -> bool

pub fn set_ep9_in(&mut self, val: bool)

pub const fn ep9_out(&self) -> bool

pub fn set_ep9_out(&mut self, val: bool)

pub const fn ep10_in(&self) -> bool

pub fn set_ep10_in(&mut self, val: bool)

pub const fn ep10_out(&self) -> bool

pub fn set_ep10_out(&mut self, val: bool)

pub const fn ep11_in(&self) -> bool

pub fn set_ep11_in(&mut self, val: bool)

pub const fn ep11_out(&self) -> bool

pub fn set_ep11_out(&mut self, val: bool)

pub const fn ep12_in(&self) -> bool

pub fn set_ep12_in(&mut self, val: bool)

pub const fn ep12_out(&self) -> bool

pub fn set_ep12_out(&mut self, val: bool)

pub const fn ep13_in(&self) -> bool

pub fn set_ep13_in(&mut self, val: bool)

pub const fn ep13_out(&self) -> bool

pub fn set_ep13_out(&mut self, val: bool)

pub const fn ep14_in(&self) -> bool

pub fn set_ep14_in(&mut self, val: bool)

pub const fn ep14_out(&self) -> bool

pub fn set_ep14_out(&mut self, val: bool)

pub const fn ep15_in(&self) -> bool

pub fn set_ep15_in(&mut self, val: bool)

pub const fn ep15_out(&self) -> bool

pub fn set_ep15_out(&mut self, val: bool)