Struct rp2040_pac2::usb::regs::Int

#[repr(transparent)]
pub struct Int(pub u32);

Interrupt Force

Tuple Fields
0: u32
Implementations

impl Int

pub const fn host_conn_dis(&self) -> bool

Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

pub fn set_host_conn_dis(&mut self, val: bool)

Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

pub const fn host_resume(&self) -> bool

Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

pub fn set_host_resume(&mut self, val: bool)

Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

pub const fn host_sof(&self) -> bool

Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

pub fn set_host_sof(&mut self, val: bool)

Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

pub const fn trans_complete(&self) -> bool

Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

pub fn set_trans_complete(&mut self, val: bool)

Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

pub const fn buff_status(&self) -> bool

Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

pub fn set_buff_status(&mut self, val: bool)

Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

pub const fn error_data_seq(&self) -> bool

Source: SIE_STATUS.DATA_SEQ_ERROR

pub fn set_error_data_seq(&mut self, val: bool)

Source: SIE_STATUS.DATA_SEQ_ERROR

pub const fn error_rx_timeout(&self) -> bool

Source: SIE_STATUS.RX_TIMEOUT

pub fn set_error_rx_timeout(&mut self, val: bool)

Source: SIE_STATUS.RX_TIMEOUT

pub const fn error_rx_overflow(&self) -> bool

Source: SIE_STATUS.RX_OVERFLOW

pub fn set_error_rx_overflow(&mut self, val: bool)

Source: SIE_STATUS.RX_OVERFLOW

pub const fn error_bit_stuff(&self) -> bool

Source: SIE_STATUS.BIT_STUFF_ERROR

pub fn set_error_bit_stuff(&mut self, val: bool)

Source: SIE_STATUS.BIT_STUFF_ERROR

pub const fn error_crc(&self) -> bool

Source: SIE_STATUS.CRC_ERROR

pub fn set_error_crc(&mut self, val: bool)

Source: SIE_STATUS.CRC_ERROR

pub const fn stall(&self) -> bool

Source: SIE_STATUS.STALL_REC

pub fn set_stall(&mut self, val: bool)

Source: SIE_STATUS.STALL_REC

pub const fn vbus_detect(&self) -> bool

Source: SIE_STATUS.VBUS_DETECT

pub fn set_vbus_detect(&mut self, val: bool)

Source: SIE_STATUS.VBUS_DETECT

pub const fn bus_reset(&self) -> bool

Source: SIE_STATUS.BUS_RESET

pub fn set_bus_reset(&mut self, val: bool)

Source: SIE_STATUS.BUS_RESET

pub const fn dev_conn_dis(&self) -> bool

Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

pub fn set_dev_conn_dis(&mut self, val: bool)

Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

pub const fn dev_suspend(&self) -> bool

Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

pub fn set_dev_suspend(&mut self, val: bool)

Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

pub const fn dev_resume_from_host(&self) -> bool

Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

pub fn set_dev_resume_from_host(&mut self, val: bool)

Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

pub const fn setup_req(&self) -> bool

Device. Source: SIE_STATUS.SETUP_REC

pub fn set_setup_req(&mut self, val: bool)

Device. Source: SIE_STATUS.SETUP_REC

pub const fn dev_sof(&self) -> bool

Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

pub fn set_dev_sof(&mut self, val: bool)

Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

pub const fn abort_done(&self) -> bool

Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

pub fn set_abort_done(&mut self, val: bool)

Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

pub const fn ep_stall_nak(&self) -> bool

Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.

pub fn set_ep_stall_nak(&mut self, val: bool)

Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.