Struct rp2040_pac2::usb::regs::SieStatus

#[repr(transparent)]pub struct SieStatus(pub u32);

SIE status register

Implementations

impl SieStatus

pub const fn data_seq_error(&self) -> bool

Data Sequence Error. The device can raise a sequence error in the following conditions: * A SETUP packet is received followed by a DATA1 packet (data phase should always be DATA0) * An OUT packet is received from the host but doesn’t match the data pid in the buffer control register read from DPSRAM The host can raise a data sequence error in the following conditions: * An IN packet from the device has the wrong data PID

pub fn set_data_seq_error(&mut self, val: bool)

Data Sequence Error. The device can raise a sequence error in the following conditions: * A SETUP packet is received followed by a DATA1 packet (data phase should always be DATA0) * An OUT packet is received from the host but doesn’t match the data pid in the buffer control register read from DPSRAM The host can raise a data sequence error in the following conditions: * An IN packet from the device has the wrong data PID

pub const fn ack_rec(&self) -> bool

ACK received. Raised by both host and device.

pub fn set_ack_rec(&mut self, val: bool)

ACK received. Raised by both host and device.

pub const fn stall_rec(&self) -> bool

Host: STALL received

pub fn set_stall_rec(&mut self, val: bool)

Host: STALL received

pub const fn nak_rec(&self) -> bool

Host: NAK received

pub fn set_nak_rec(&mut self, val: bool)

Host: NAK received

pub const fn rx_timeout(&self) -> bool

RX timeout is raised by both the host and device if an ACK is not received in the maximum time specified by the USB spec.

pub fn set_rx_timeout(&mut self, val: bool)

RX timeout is raised by both the host and device if an ACK is not received in the maximum time specified by the USB spec.

pub const fn rx_overflow(&self) -> bool

RX overflow is raised by the Serial RX engine if the incoming data is too fast.

pub fn set_rx_overflow(&mut self, val: bool)

RX overflow is raised by the Serial RX engine if the incoming data is too fast.

pub const fn bit_stuff_error(&self) -> bool

Bit Stuff Error. Raised by the Serial RX engine.

pub fn set_bit_stuff_error(&mut self, val: bool)

Bit Stuff Error. Raised by the Serial RX engine.

pub const fn crc_error(&self) -> bool

CRC Error. Raised by the Serial RX engine.

pub fn set_crc_error(&mut self, val: bool)

CRC Error. Raised by the Serial RX engine.

pub const fn bus_reset(&self) -> bool

Device: bus reset received

pub fn set_bus_reset(&mut self, val: bool)

Device: bus reset received

pub const fn trans_complete(&self) -> bool

Transaction complete. Raised by device if: * An IN or OUT packet is sent with the LAST_BUFF bit set in the buffer control register Raised by host if: * A setup packet is sent when no data in or data out transaction follows * An IN packet is received and the LAST_BUFF bit is set in the buffer control register * An IN packet is received with zero length * An OUT packet is sent and the LAST_BUFF bit is set

pub fn set_trans_complete(&mut self, val: bool)

Transaction complete. Raised by device if: * An IN or OUT packet is sent with the LAST_BUFF bit set in the buffer control register Raised by host if: * A setup packet is sent when no data in or data out transaction follows * An IN packet is received and the LAST_BUFF bit is set in the buffer control register * An IN packet is received with zero length * An OUT packet is sent and the LAST_BUFF bit is set

pub const fn setup_rec(&self) -> bool

Device: Setup packet received

pub fn set_setup_rec(&mut self, val: bool)

Device: Setup packet received

pub const fn connected(&self) -> bool

Device: connected

pub fn set_connected(&mut self, val: bool)

Device: connected

pub const fn resume(&self) -> bool

Host: Device has initiated a remote resume. Device: host has initiated a resume.

pub fn set_resume(&mut self, val: bool)

Host: Device has initiated a remote resume. Device: host has initiated a resume.

pub const fn vbus_over_curr(&self) -> bool

VBUS over current detected

pub fn set_vbus_over_curr(&mut self, val: bool)

VBUS over current detected

pub const fn speed(&self) -> u8

Host: device speed. Disconnected = 00, LS = 01, FS = 10

pub fn set_speed(&mut self, val: u8)

Host: device speed. Disconnected = 00, LS = 01, FS = 10

pub const fn suspended(&self) -> bool

Bus in suspended state. Valid for device and host. Host and device will go into suspend if neither Keep Alive / SOF frames are enabled.

pub fn set_suspended(&mut self, val: bool)

Bus in suspended state. Valid for device and host. Host and device will go into suspend if neither Keep Alive / SOF frames are enabled.

pub const fn line_state(&self) -> u8

USB bus line state

pub fn set_line_state(&mut self, val: u8)

USB bus line state

pub const fn vbus_detected(&self) -> bool

Device: VBUS Detected

pub fn set_vbus_detected(&mut self, val: bool)

Device: VBUS Detected