Struct rp2040_pac2::usbctrl_regs::UsbctrlRegs

pub struct UsbctrlRegs(pub *mut u8);

USB FS/LS controller device registers

Implementations

impl UsbctrlRegs

pub fn addr_endp(self) -> Reg<AddrEndp, RW>

Device address and endpoint control

pub fn addr_endp1(self) -> Reg<AddrEndp1, RW>

Interrupt endpoint 1. Only valid for HOST mode.

pub fn addr_endp2(self) -> Reg<AddrEndp2, RW>

Interrupt endpoint 2. Only valid for HOST mode.

pub fn addr_endp3(self) -> Reg<AddrEndp3, RW>

Interrupt endpoint 3. Only valid for HOST mode.

pub fn addr_endp4(self) -> Reg<AddrEndp4, RW>

Interrupt endpoint 4. Only valid for HOST mode.

pub fn addr_endp5(self) -> Reg<AddrEndp5, RW>

Interrupt endpoint 5. Only valid for HOST mode.

pub fn addr_endp6(self) -> Reg<AddrEndp6, RW>

Interrupt endpoint 6. Only valid for HOST mode.

pub fn addr_endp7(self) -> Reg<AddrEndp7, RW>

Interrupt endpoint 7. Only valid for HOST mode.

pub fn addr_endp8(self) -> Reg<AddrEndp8, RW>

Interrupt endpoint 8. Only valid for HOST mode.

pub fn addr_endp9(self) -> Reg<AddrEndp9, RW>

Interrupt endpoint 9. Only valid for HOST mode.

pub fn addr_endp10(self) -> Reg<AddrEndp10, RW>

Interrupt endpoint 10. Only valid for HOST mode.

pub fn addr_endp11(self) -> Reg<AddrEndp11, RW>

Interrupt endpoint 11. Only valid for HOST mode.

pub fn addr_endp12(self) -> Reg<AddrEndp12, RW>

Interrupt endpoint 12. Only valid for HOST mode.

pub fn addr_endp13(self) -> Reg<AddrEndp13, RW>

Interrupt endpoint 13. Only valid for HOST mode.

pub fn addr_endp14(self) -> Reg<AddrEndp14, RW>

Interrupt endpoint 14. Only valid for HOST mode.

pub fn addr_endp15(self) -> Reg<AddrEndp15, RW>

Interrupt endpoint 15. Only valid for HOST mode.

pub fn main_ctrl(self) -> Reg<MainCtrl, RW>

Main control register

pub fn sof_wr(self) -> Reg<SofWr, RW>

Set the SOF (Start of Frame) frame number in the host controller. The SOF packet is sent every 1ms and the host will increment the frame number by 1 each time.

pub fn sof_rd(self) -> Reg<SofRd, RW>

Read the last SOF (Start of Frame) frame number seen. In device mode the last SOF received from the host. In host mode the last SOF sent by the host.

pub fn sie_ctrl(self) -> Reg<SieCtrl, RW>

SIE control register

pub fn sie_status(self) -> Reg<SieStatus, RW>

SIE status register

pub fn int_ep_ctrl(self) -> Reg<IntEpCtrl, RW>

interrupt endpoint control register

pub fn buff_status(self) -> Reg<BuffStatus, RW>

Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

pub fn buff_cpu_should_handle(self) -> Reg<BuffCpuShouldHandle, RW>

Which of the double buffers should be handled. Only valid if using an interrupt per buffer (i.e. not per 2 buffers). Not valid for host interrupt endpoint polling because they are only single buffered.

pub fn ep_abort(self) -> Reg<EpAbort, RW>

Device only: Can be set to ignore the buffer control register for this endpoint in case you would like to revoke a buffer. A NAK will be sent for every access to the endpoint until this bit is cleared. A corresponding bit in EP_ABORT_DONE is set when it is safe to modify the buffer control register.

pub fn ep_abort_done(self) -> Reg<EpAbortDone, RW>

Device only: Used in conjunction with EP_ABORT. Set once an endpoint is idle so the programmer knows it is safe to modify the buffer control register.

pub fn ep_stall_arm(self) -> Reg<EpStallArm, RW>

Device: this bit must be set in conjunction with the STALL bit in the buffer control register to send a STALL on EP0. The device controller clears these bits when a SETUP packet is received because the USB spec requires that a STALL condition is cleared when a SETUP packet is received.

pub fn nak_poll(self) -> Reg<NakPoll, RW>

Used by the host controller. Sets the wait time in microseconds before trying again if the device replies with a NAK.

pub fn ep_status_stall_nak(self) -> Reg<EpStatusStallNak, RW>

Device: bits are set when the IRQ_ON_NAK or IRQ_ON_STALL bits are set. For EP0 this comes from SIE_CTRL. For all other endpoints it comes from the endpoint control register.

pub fn usb_muxing(self) -> Reg<UsbMuxing, RW>

Where to connect the USB controller. Should be to_phy by default.

pub fn usb_pwr(self) -> Reg<UsbPwr, RW>

Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

pub fn usbphy_direct(self) -> Reg<UsbphyDirect, RW>

This register allows for direct control of the USB phy. Use in conjunction with usbphy_direct_override register to enable each override bit.

pub fn usbphy_direct_override(self) -> Reg<UsbphyDirectOverride, RW>

Override enable for each control in usbphy_direct

pub fn usbphy_trim(self) -> Reg<UsbphyTrim, RW>

Used to adjust trim values of USB phy pull down resistors.

pub fn intr(self) -> Reg<Intr, RW>

Raw Interrupts

pub fn inte(self) -> Reg<Inte, RW>

Interrupt Enable

pub fn intf(self) -> Reg<Intf, RW>

Interrupt Force

pub fn ints(self) -> Reg<Ints, RW>

Interrupt status after masking & forcing