Struct rp2040_pac2::adc::regs::Cs

#[repr(transparent)]
pub struct Cs(pub u32);

ADC Control and Status

Tuple Fields
0: u32
Implementations

impl Cs

pub const fn en(&self) -> bool

Power on ADC and enable its clock. 1 - enabled. 0 - disabled.

pub fn set_en(&mut self, val: bool)

Power on ADC and enable its clock. 1 - enabled. 0 - disabled.

pub const fn ts_en(&self) -> bool

Power on temperature sensor. 1 - enabled. 0 - disabled.

pub fn set_ts_en(&mut self, val: bool)

Power on temperature sensor. 1 - enabled. 0 - disabled.

pub const fn start_once(&self) -> bool

Start a single conversion. Self-clearing. Ignored if start_many is asserted.

pub fn set_start_once(&mut self, val: bool)

Start a single conversion. Self-clearing. Ignored if start_many is asserted.

pub const fn start_many(&self) -> bool

Continuously perform conversions whilst this bit is 1. A new conversion will start immediately after the previous finishes.

pub fn set_start_many(&mut self, val: bool)

Continuously perform conversions whilst this bit is 1. A new conversion will start immediately after the previous finishes.

pub const fn ready(&self) -> bool

1 if the ADC is ready to start a new conversion. Implies any previous conversion has completed. 0 whilst conversion in progress.

pub fn set_ready(&mut self, val: bool)

1 if the ADC is ready to start a new conversion. Implies any previous conversion has completed. 0 whilst conversion in progress.

pub const fn err(&self) -> bool

The most recent ADC conversion encountered an error; result is undefined or noisy.

pub fn set_err(&mut self, val: bool)

The most recent ADC conversion encountered an error; result is undefined or noisy.

pub const fn err_sticky(&self) -> bool

Some past ADC conversion encountered an error. Write 1 to clear.

pub fn set_err_sticky(&mut self, val: bool)

Some past ADC conversion encountered an error. Write 1 to clear.

pub const fn ainsel(&self) -> u8

Select analog mux input. Updated automatically in round-robin mode.

pub fn set_ainsel(&mut self, val: u8)

Select analog mux input. Updated automatically in round-robin mode.

pub const fn rrobin(&self) -> u8

Round-robin sampling. 1 bit per channel. Set all bits to 0 to disable. Otherwise, the ADC will cycle through each enabled channel in a round-robin fashion. The first channel to be sampled will be the one currently indicated by AINSEL. AINSEL will be updated after each conversion with the newly-selected channel.

pub fn set_rrobin(&mut self, val: u8)

Round-robin sampling. 1 bit per channel. Set all bits to 0 to disable. Otherwise, the ADC will cycle through each enabled channel in a round-robin fashion. The first channel to be sampled will be the one currently indicated by AINSEL. AINSEL will be updated after each conversion with the newly-selected channel.